remove GsltDelete, move delete to credentials
