+++ src/gslt/mod.rs
@@ -4,7 +4,6 @@ use url::form_urlencoded;
 
 pub mod credential;
 pub mod list;
-pub mod delete;
 pub mod setmemo;
 
 use credential::GsltCredentialResponse;
@@ -41,9 +40,6 @@ pub trait GsltList {
 
 pub trait GsltCredential {
     fn create_gslt(&self, app_id: u32, memo: impl Into<String>) -> Result<GsltCredentialResponse>;
-}
-
-pub trait GsltDelete {
     fn delete_gslt(&self, steam_id: u64) -> Result<()>;
 }
